+++ cardgamesgui/src/main/java/cards/gui/dialogs/EditorBelote.java
@@ -325,7 +325,7 @@ public final class EditorBelote extends DialogBelote implements SetterSelectedCa
         byte donneur_ = (byte) liste.getSelectedIndex();
         if (donneur_ == nombreDeJoueurs_) {
 //          donneur_=(byte)Math.floor(nombreDeJoueurs_*MonteCarlo.randomDouble());
-            donneur_=(byte)AbMonteCarlo.randomInt(nombreDeJoueurs_);
+            donneur_=(byte)AbMonteCarlo.randomLong(nombreDeJoueurs_);
         }
         DealBelote donne_=new DealBelote(mains_,donneur_);
         partie = new GameBelote(GameType.EDIT,donne_,getReglesBelote());
